Source char_traits<unsigned char> from GCC C++ Library directly to clarify license.
http://gcc.gnu.org/onlinedocs/libstdc++/
Previously borrowed code from id3lib was considered shady.
Placed in its own file for style reasons.
